Move more stuff into hang.
